Improved messaging Data files moved states (will eventually add to git ignore)
