+++ text_posts/posted/0.txt
@@ -1,1 +1,1 @@
-This is the zeroth text post, not in order alphabetically but in order of creation
+Tweet 1 in a series #TestingAutomation
