+++ text_posts/posted/1.txt
@@ -1,1 +1,1 @@
-This is the first text post
+Second Tweet!
