+++ text_posts/posted/3.txt
@@ -1,1 +1,1 @@
-This is the third text post
+FOUR FOUR FOUR
