+++ text_posts/posted/4.txt
@@ -1,0 +1,1 @@
+F I V E
