+++ text_posts/posted/5.txt
@@ -1,0 +1,1 @@
+666 RONALD REAGAN 666
